Ditch scatterplot superclass for now; easy to recreate
