+++ src/main/java/org/veupathdb/service/eda/ds/plugin/abundance/AbundanceScatterplotPlugin.java
@@ -5,11 +5,13 @@ import org.apache.logging.log4j.Logger;
 import org.gusdb.fgputil.ListBuilder;
 import org.gusdb.fgputil.validation.ValidationException;
 import org.veupathdb.service.eda.common.client.spec.StreamSpec;
+import org.veupathdb.service.eda.common.plugin.constraint.ConstraintSpec;
 import org.veupathdb.service.eda.common.plugin.constraint.DataElementSet;
 import org.veupathdb.service.eda.common.plugin.util.PluginUtil;
 import org.veupathdb.service.eda.common.plugin.util.RServeClient;
 import org.veupathdb.service.eda.ds.Resources;
-import org.veupathdb.service.eda.ds.plugin.AbstractScatterplotWithCompute;
+import org.veupathdb.service.eda.ds.metadata.AppsMetadata;
+import org.veupathdb.service.eda.ds.plugin.AbstractPlugin;
 import org.veupathdb.service.eda.generated.model.*;
 
 import java.io.IOException;
@@ -22,9 +24,14 @@ import java.util.Map;
 
 import static org.veupathdb.service.eda.common.plugin.util.RServeClient.useRConnectionWithRemoteFiles;
 
-public class AbundanceScatterplotPlugin extends AbstractScatterplotWithCompute<AbundanceScatterplotPostRequest, ScatterplotWith1ComputeSpec, RankedAbundanceComputeConfig> {
+public class AbundanceScatterplotPlugin extends AbstractPlugin<AbundanceScatterplotPostRequest, ScatterplotWith1ComputeSpec, RankedAbundanceComputeConfig> {
 
   private static final Logger LOG = LogManager.getLogger(AbundanceScatterplotPlugin.class);
+  
+  @Override
+  public String getDisplayName() {
+    return "Scatter plot";
+  }
 
   @Override
   public String getDescription() {
@@ -32,10 +39,31 @@ public class AbundanceScatterplotPlugin extends AbstractScatterplotWithCompute<A
   }
 
   @Override
+  public List<String> getProjects() {
+    return List.of(AppsMetadata.MICROBIOME_PROJECT);
+  }
+
+  @Override
   protected ClassGroup getTypeParameterClasses() {
     return new ClassGroup(AbundanceScatterplotPostRequest.class, ScatterplotWith1ComputeSpec.class, RankedAbundanceComputeConfig.class);
   }
 
+  @Override
+  public ConstraintSpec getConstraintSpec() {
+    return new ConstraintSpec()
+      .dependencyOrder(List.of("xAxisVariable"), List.of("facetVariable"))
+      .pattern()
+        .element("xAxisVariable")
+          .types(APIVariableType.NUMBER, APIVariableType.DATE, APIVariableType.INTEGER)
+          .description("Variable must be a number or date and be of the same or a parent entity as the Y-axis variable.")
+        .element("facetVariable")
+          .required(false)
+          .maxVars(2)
+          .maxValues(7)
+          .description("Variable(s) must have7 or fewer unique values and be of the same or a parent entity as the Overlay variable.")
+      .done();
+  }
+  
   @Override
   protected void validateVisualizationSpec(ScatterplotWith1ComputeSpec pluginSpec) throws ValidationException {
     validateInputs(new DataElementSet()
