+++ src/main/java/org/veupathdb/service/eda/ds/plugin/alphadiv/AlphaDivScatterplotPlugin.java
@@ -4,12 +4,13 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.gusdb.fgputil.validation.ValidationException;
 import org.veupathdb.service.eda.common.client.spec.StreamSpec;
+import org.veupathdb.service.eda.common.plugin.constraint.ConstraintSpec;
 import org.veupathdb.service.eda.common.plugin.constraint.DataElementSet;
 import org.veupathdb.service.eda.common.plugin.util.PluginUtil;
 import org.veupathdb.service.eda.common.plugin.util.RServeClient;
 import org.veupathdb.service.eda.ds.Resources;
 import org.veupathdb.service.eda.ds.metadata.AppsMetadata;
-import org.veupathdb.service.eda.ds.plugin.AbstractScatterplotWithCompute;
+import org.veupathdb.service.eda.ds.plugin.AbstractPlugin;
 import org.veupathdb.service.eda.generated.model.*;
 
 import java.io.IOException;
@@ -21,7 +22,7 @@ import java.util.Map;
 
 import static org.veupathdb.service.eda.common.plugin.util.RServeClient.useRConnectionWithRemoteFiles;
 
-public class AlphaDivScatterplotPlugin extends AbstractScatterplotWithCompute<AlphaDivScatterplotPostRequest, ScatterplotWith1ComputeSpec, AlphaDivComputeConfig> {
+public class AlphaDivScatterplotPlugin extends AbstractPlugin<AlphaDivScatterplotPostRequest, ScatterplotWith1ComputeSpec, AlphaDivComputeConfig> {
 
   private static final Logger LOG = LogManager.getLogger(AlphaDivScatterplotPlugin.class);
   
@@ -45,6 +46,39 @@ public class AlphaDivScatterplotPlugin extends AbstractScatterplotWithCompute<Al
     return new ClassGroup(AlphaDivScatterplotPostRequest.class, ScatterplotWith1ComputeSpec.class, AlphaDivComputeConfig.class);
   }
 
+  @Override
+  public ConstraintSpec getConstraintSpec() {
+    return new ConstraintSpec()
+      .dependencyOrder(List.of("xAxisVariable"), List.of("overlayVariable", "facetVariable"))
+      .pattern()
+        .element("xAxisVariable")
+          .types(APIVariableType.NUMBER, APIVariableType.DATE, APIVariableType.INTEGER)
+          .description("Variable must be a number or date and be of the same or a parent entity as the Y-axis variable.")
+        .element("overlayVariable")
+          .required(false)
+          .maxValues(8)
+          .description("Variable must be a number, or have 8 or fewer values, and be of the same or a parent entity as the X-axis variable.")
+        .element("facetVariable")
+          .required(false)
+          .maxVars(2)
+          .maxValues(10)
+          .description("Variable(s) must have 10 or fewer unique values and be of the same or a parent entity as the Overlay variable.")
+      .pattern()
+        .element("xAxisVariable")
+          .types(APIVariableType.NUMBER, APIVariableType.DATE, APIVariableType.INTEGER)
+          .description("Variable must be a number or date and be of the same or a parent entity as the Y-axis variable.")
+        .element("overlayVariable")
+          .required(false)
+          .types(APIVariableType.NUMBER, APIVariableType.INTEGER) 
+          .description("Variable must be a number, or have 8 or fewer values, and be of the same or a parent entity as the X-axis variable.")
+        .element("facetVariable")
+          .required(false)
+          .maxVars(2)
+          .maxValues(10)
+          .description("Variable(s) must have 10 or fewer unique values and be of the same or a parent entity as the Overlay variable.")
+      .done();
+  }
+  
   @Override
   protected void validateVisualizationSpec(ScatterplotWith1ComputeSpec pluginSpec) throws ValidationException {
     validateInputs(new DataElementSet()
