+++ src/main/java/org/veupathdb/service/eda/ds/plugin/betadiv/BetaDivScatterplotPlugin.java
@@ -1,36 +1,64 @@
 package org.veupathdb.service.eda.ds.plugin.betadiv;
 
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-import org.gusdb.fgputil.validation.ValidationException;
-import org.veupathdb.service.eda.common.client.spec.StreamSpec;
-import org.veupathdb.service.eda.common.plugin.constraint.DataElementSet;
-import org.veupathdb.service.eda.common.plugin.util.RServeClient;
-import org.veupathdb.service.eda.ds.Resources;
-import org.veupathdb.service.eda.ds.plugin.AbstractScatterplotWithCompute;
-import org.veupathdb.service.eda.generated.model.*;
-
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+import org.gusdb.fgputil.validation.ValidationException;
+import org.veupathdb.service.eda.common.client.spec.StreamSpec;
+import org.veupathdb.service.eda.common.plugin.constraint.ConstraintSpec;
+import org.veupathdb.service.eda.common.plugin.constraint.DataElementSet;
+import org.veupathdb.service.eda.ds.Resources;
+import org.veupathdb.service.eda.ds.metadata.AppsMetadata;
+import org.veupathdb.service.eda.common.plugin.util.RServeClient;
+import org.veupathdb.service.eda.ds.plugin.AbstractPlugin;
+import org.veupathdb.service.eda.generated.model.*;
 
 import static org.veupathdb.service.eda.common.plugin.util.RServeClient.useRConnectionWithRemoteFiles;
 
-public class BetaDivScatterplotPlugin extends AbstractScatterplotWithCompute<BetaDivScatterplotPostRequest, BetaDivScatterplotSpec, BetaDivComputeConfig> {
+public class BetaDivScatterplotPlugin extends AbstractPlugin<BetaDivScatterplotPostRequest, BetaDivScatterplotSpec, BetaDivComputeConfig> {
 
   private static final Logger LOG = LogManager.getLogger(BetaDivScatterplotPlugin.class);
+  
+  @Override
+  public String getDisplayName() {
+    return "Scatter plot";
+  }
 
   @Override
   public String getDescription() {
-    return "Visualize a 2-dimensional projection of samples based on their beta diversity";
+    return "Visualize a 2-dimensional projection of samples based on their beta diversitiy";
+  }
+
+  @Override
+  public List<String> getProjects() {
+    return List.of(AppsMetadata.MICROBIOME_PROJECT);
   }
 
   @Override
   protected ClassGroup getTypeParameterClasses() {
     return new ClassGroup(BetaDivScatterplotPostRequest.class, BetaDivScatterplotSpec.class, BetaDivComputeConfig.class);
+  }
+
+  @Override
+  public ConstraintSpec getConstraintSpec() {
+    return new ConstraintSpec()
+      .dependencyOrder(List.of("overlayVariable"))
+      .pattern()
+        .element("overlayVariable")
+          .required(false)
+          .maxValues(8)
+          .description("Variable must be a number, or have 8 or fewer values, and be of the same or a parent entity as the X-axis variable.")
+      .pattern()
+        .element("overlayVariable")
+          .required(false)
+          .types(APIVariableType.NUMBER, APIVariableType.INTEGER) 
+          .description("Variable must be a number, or have 8 or fewer values, and be of the same or a parent entity as the X-axis variable.")
+      .done();
   }
 
   @Override
